requirements.txt
# oTree-may-not-overwrite-this-file
otree[mturk]==5.3.1
psycopg2>=2.8.4
sentry-sdk>=0.7.9
Pillow>=8.3.2
numpy>=1.3.2
